+++ src/icons/music-audio/index.ts
@@ -1,0 +1,3 @@
+import Mic from '@/src/icons/music-audio/Mic';
+
+export { Mic };