+++ src/icons/nature-environment/index.ts
@@ -1,0 +1,3 @@
+import Fire from '@/src/icons/nature-environment/Fire';
+
+export { Fire };