+++ src/icons/political/index.ts
@@ -1,0 +1,3 @@
+import Flag from '@/src/icons/political/Flag';
+
+export { Flag };