+++ src/icons/shopping/index.ts
@@ -1,6 +1,7 @@
 /**
  * This file exports all icons in the "Shopping" category.
  */
+import Discount from '@/src/icons/shopping/Discount';
 import ShoppingCart from '@/src/icons/shopping/ShoppingCart';
 
-export { ShoppingCart };
+export { Discount, ShoppingCart };